New View and controller for Print feature
- Created a new print.html and PrintController.js files
- The new view will load every call into the same view at once and after
the data is ready it will call for printing the view. I'm passing the
zoom for the map, payment type, saving method, and shipping type n in
the url e.g:

http://ec2-54-84-156-215.compute-1.amazonaws.com/home/#/print/266?activeView=standardShipping&paymentType=leaseToOwn&savingsMethod=immediate&mapZoom=19

- I had to rewrite most of the styling because most of it did not fit in
the printing view. Also, in order to view the background colors you will
have to select the background graphics option in the browser, if you
don't it won't show the greens in the totals...

- Right now the view fully works in portrait mode, if there is a need
for landscape or other type of formats that will take some more time and
we would need to pass that in the URL as well

- I put the brackets table and the existing and proposed light fixtures
tables in one page by themselves thinking that they may grow depending
on the project. If there is a need to rearrange things into different
pages let me know